featuretests: refactor route feature tests into weights tests
Refactors the tests dealing with route weights into it's own
test file.
